encode / decode Links with identity Multibase prefix
According to the spec (https://ipld.io/specs/codecs/dag-cbor/spec/#links),
CIDs MUST be encoded including the identity Multibase prefix (0x00).

This change adds the Multibase prefix in encoding (in order to follow
the specs) and requires the Multibase prefix in decoding (in order not
to become ambiguous and be in line with the spec).